+++ lib/esm/transformations/VariationGenerator.d.ts
@@ -1,0 +1,11 @@
+export function generate(config?: {
+    generateVariationPrompt: string;
+    noOfVariations: number;
+    seed: number;
+    autoscale: boolean;
+}): Transformation;
+declare namespace _default {
+    export { generate };
+}
+export default _default;
+import Transformation from "../transformation.js";